third_party/blink/web_tests/html/document_metadata/referrer-inside-head-not-use-counted.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<head>
  <meta name=referrer content="origin, origin" />
</head>
<script>
  'use strict';

  test(() => {
    const kHTMLMetaElementReferrerPolicyOutsideHead = 3314; // from web_feature.mojom
    assert_false(internals.isUseCounted(document, kHTMLMetaElementReferrerPolicyOutsideHead));
  }, '<meta name=referrer> inside <head> is not use counted');
</script>
